Allow Restrictions to be defined and discarded
Restrictions are persisted when user selects Done button. If the Restrictions have changed then existing Assignments are deleted otherwise any existing Assignments are unaffected.

+++ open-secret-santa-app/src/main/java/com/moac/android/opensecretsanta/activity/NewDrawActivity.java
@@ -149,7 +149,7 @@ public class NewDrawActivity extends Activity implements DrawManager {
     }
 
     private void prepareDraw(Group _group) {
-        int count = mDb.deleteAllAssignmentsForGroup(_group.getId());
+        long count = mDb.deleteAllAssignmentsForGroup(_group.getId());
         Log.v(TAG, "prepareDraw() - deleted Assignment count: " + count);
         mMembersListFragment.onAssignmentsCleared();
     }
